requirements.txt
fastapi==0.120.4
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
pydantic==2.12.3
pydantic-settings==2.11.0
pillow==12.0.0
pytesseract==0.3.13
psycopg2-binary==2.9.11
python-multipart==0.0.20
requests==2.32.5
PyMuPDF==1.23.8

# Optional OCR providers (uncomment if needed)
# google-cloud-vision==3.4.0
# azure-cognitiveservices-vision-computervision==0.9.0
# msrest==0.7.1

